Tolerate new udev actions "bind" and "unbind"
Recent kernel and udev versions have added new actions that are triggered on top of the usual add and remove actions. This causes the script to `exit 1` every time, which appears as a warning in udev logs. Why not a big deal, this is ugly, this patch fixes that and protects us against future new events as well.